MDL-44838 Area:fix ical import wholeday events
The current issue was that whole-day events in ical files were not correctly
imported into moodlecalendar. For a whole-day event, usually there is no timezone
set in the ical file and there are only dates set, an ical file for a whole day event
 for example looks like this
DTSTART;VALUE=DATE:20241203
DTEND;VALUE=DATE:20241204

Since there are no times set, moodle interprets the date as
DTSTART;VALUE=DATE:20241203T000000
DTEND;VALUE=DATE:20241204T000000
As a consequence the one day event is marked in the calendar as a two day event,
 which ends a 0:00am on the second day. This can be very confusing.
(e.g. holidays are marked one additional day longer as they are supposed to!)

To mitigate this problem, every imported event is checked, if the value DTSTART has a time (hour).
If it doesn’t, it is a whole day event without specified time.
Then the timezone is set to local server time, and 1 minute is subtracted from the DTEND.
 This way, the calendar is marked correctly in the moodle as a one day event from 0:00:00 to 23:59:00.
An ical export of the imported event from moodle would yield
DTSTART;VALUE=DATE:20241203T000000
DTEND;VALUE=DATE:20241203T235900